(svn r25117) -Fix [FS#5511]: Silence warning.

+++ src/game/game_text.cpp
@@ -263,7 +263,7 @@ GameStrings *LoadTranslations()
 		/* Scan for other language files */
 		LanguageScanner scanner(gs, filename);
 		strecpy(e, "lang" PATHSEP, lastof(filename));
-		uint len = strlen(filename);
+		size_t len = strlen(filename);
 
 		const char *tar_filename = info->GetTarFile();
 		TarList::iterator iter;
